Move Location ORM object to separate model file and add get_all_locations and serialize methods
