+++ models/__init__.py
@@ -1,0 +1,5 @@
+from models.location import Location
+
+
+# List all the imports to comply with pep8
+__all__ = [Location]
